scap next related functions completely inlined refactor of scap-int to avoid zlib dependencies in sinsp added calls for thread cache reset after thread removal added calls to init state of filtering engine in scap added calls to enable filtering in sinsp added counters for file filtering and exit events filtering
